refactor updating mail imports
