+++ apps/bilan-carbone/src/app/(dashboard)/layout.tsx
@@ -2,7 +2,6 @@ import ChecklistButton from '@/components/checklist/ChecklistButton'
 import withAuth, { UserSessionProps } from '@/components/hoc/withAuth'
 import Navbar from '@/components/navbar/Navbar'
 import OrganizationCard from '@/components/organizationCard/OrganizationCard'
-import { environmentsWithChecklist } from '@/constants/environments'
 import { getAccountOrganizationVersions } from '@/db/account'
 import { getAllowedStudyIdByAccount } from '@/db/study'
 import EnvironmentInitializer from '@/environments/core/EnvironmentInitializer'
@@ -11,6 +10,7 @@ import { getEnvironment } from '@/i18n/environment'
 import { isTiltSimplifiedFeatureActive } from '@/services/permissions/environment'
 import { shouldRenewLicenceText } from '@/utils/organization'
 import { Environment } from '@abc-transitionbascarbone/db-common/enums'
+import { environmentsWithChecklist } from '@abc-transitionbascarbone/utils/environments'
 import { Box } from '@mui/material'
 import classNames from 'classnames'
 import styles from './layout.module.css'
