+++ apps/bilan-carbone/src/app/(dashboard)/page.tsx
@@ -3,13 +3,13 @@ import withAuth, { UserSessionProps } from '@/components/hoc/withAuth'
 import UserFeedback from '@/components/home/UserFeedback'
 import UserView from '@/components/home/UserView'
 import Onboarding from '@/components/onboarding/Onboarding'
-import { environmentWithOnboarding } from '@/constants/environments'
 import { getOrganizationVersionById } from '@/db/organization'
 import DynamicComponent from '@/environments/core/utils/DynamicComponent'
 import { default as CUTLogosHome } from '@/environments/cut/home/LogosHome'
 import { displayFeedBackForm } from '@/services/serverFunctions/user'
 import { Environment } from '@abc-transitionbascarbone/db-common/enums'
 import { UserSession } from 'next-auth'
+import { environmentWithOnboarding } from '@abc-transitionbascarbone/utils/environments'
 import dynamic from 'next/dynamic'
 
 const ClicksonUserView = dynamic(() => import('@/environments/clickson/home/UserView'))
