+++ apps/bilan-carbone/src/components/pages/SelectAccount.tsx
@@ -1,11 +1,11 @@
 'use client'
 
-import { EnvironmentNames } from '@/constants/environments'
 import { getUserWithAccountsAndOrganizationsById } from '@/db/user'
 import { switchEnvironment } from '@/i18n/environment'
 import { accountHandler } from '@/services/auth.utils'
 import { useAppEnvironmentStore } from '@/store/AppEnvironment'
 import { Environment } from '@abc-transitionbascarbone/db-common/enums'
+import { EnvironmentNames } from '@abc-transitionbascarbone/utils/environments'
 import PermIdentityIcon from '@mui/icons-material/PermIdentity'
 import { Chip, List, ListItem, ListItemButton, ListItemIcon, ListItemText } from '@mui/material'
 import { UserSession } from 'next-auth'
