+++ apps/bilan-carbone/src/components/pages/StudyPostsContainer.tsx
@@ -1,11 +1,11 @@
 'use client'
-import { EnvironmentMode } from '@/constants/environments'
 import type { FullStudy } from '@/db/study'
 import DynamicComponent from '@/environments/core/utils/DynamicComponent'
 import { customRich } from '@/i18n/customRich'
 import { Post, subPostsByPost } from '@/services/posts'
 import { SimplifiedEnvironment } from '@/services/publicodes/simplifiedPublicodesConfig'
 import { StudyRole, SubPost } from '@abc-transitionbascarbone/db-common/enums'
+import { EnvironmentMode } from '@abc-transitionbascarbone/utils/environments'
 import { CircularProgress } from '@mui/material'
 import { UserSession } from 'next-auth'
 import { useTranslations } from 'next-intl'
