+++ apps/bilan-carbone/src/environments/core/utils/DynamicComponent.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import { EnvironmentMode } from '@/constants/environments'
 import { isAdvanced, isSimplified } from '@/services/permissions/environment'
 import { useAppEnvironmentStore } from '@/store/AppEnvironment'
 import { Environment } from '@abc-transitionbascarbone/db-common/enums'
+import { EnvironmentMode } from '@abc-transitionbascarbone/utils/environments'
 import { ReactNode } from 'react'
 import EnvironmentLoader from './EnvironmentLoader'
