+++ apps/bilan-carbone/src/utils/ressources.ts
@@ -1,7 +1,7 @@
-import { getEnvVar } from '@/lib/environment'
 import { Environment } from '@abc-transitionbascarbone/db-common/enums'
 import { Locale } from '@abc-transitionbascarbone/i18n/config'
 import { Translations } from '@abc-transitionbascarbone/lib'
+import { getEnvVar } from '@abc-transitionbascarbone/lib/environment'
 import { getLocale } from 'next-intl/server'
 
 export const getEnvironnementRessources = async (env: Environment, t: Translations) => {
