+++ packages/services/package.json
@@ -7,5 +7,9 @@
   "exports": {
     ".": "./src/index.ts",
     "./*": "./src/*.ts"
+  },
+  "dependencies": {
+    "@abc-transitionbascarbone/lib": "*",
+    "@abc-transitionbascarbone/utils": "*"
   }
 }
