+++ packages/services/src/email/send.ts
@@ -1,5 +1,5 @@
-import { getEnvVar } from '@/lib/environment'
 import { Environment } from '@abc-transitionbascarbone/db-common/enums'
+import { getEnvVar } from '@abc-transitionbascarbone/lib/environment'
 import ejs from 'ejs'
 import fs from 'fs'
 import path from 'path'
@@ -15,7 +15,7 @@ import { getTransporter } from './transposter'
  * @returns Prosmise that resolves rendered HTML string.
  */
 const getHtml = async ({ file, env, data }: { file: string; env: Environment; data?: ejs.Data }) => {
-  const basePath = path.join(process.cwd(), 'src', 'services', 'email', 'views')
+  const basePath = path.join(process.cwd(), '..', '..', 'packages', 'services', 'src', 'email', 'views')
   const customPath = path.join(basePath, env, `${file}.ejs`)
   const fallbackPath = path.join(basePath, 'common', `${file}.ejs`)
   const templatePath = fs.existsSync(customPath) ? customPath : fallbackPath
